Transfer code to new separate modules
