+++ reports-management/index.html
@@ -325,7 +325,6 @@
               </select>
             </div>
           </section>
-
           <section class="category hidden">
             <div class="form-group input-field marg3">
               <label for="reportCategory">Report category: </label>
