+++ reports-management/scripts/event-handlers/generate-event.js
