+++ reports-management/scripts/main.js
@@ -5,10 +5,11 @@
 //! For Development only.
 //! Comment out for Production.
 
-import { createReportURI } from "./button-events/generate-event.js";
+import { createReportURI } from "./event-handlers/generate-event.js";
+import { categoryHandler } from "./event-handlers/category-handler.js";
 // Shorten a list of records to a list of obj with key/value properties
 // list = [{...},{...}], keyParam, valueParam = field names as string
-const createSummaryList = (list, keyParam, valueParam) => {
+export const createSummaryList = (list, keyParam, valueParam) => {
   return list.map((record) => {
     return { key: record[keyParam], value: record[valueParam] };
   });
@@ -16,8 +17,8 @@ const createSummaryList = (list, keyParam, valueParam) => {
 //! =============================================================
 const reportGroups = GetReportCategory.slice(0);
 const reports = GetReport.slice(0);
-const prepByList = GetPrepareBy.filter((record) => record.key !== "0");
-const prepByListCM = GetPrepareByCM.slice(0);
+export const prepByList = GetPrepareBy.filter((record) => record.key !== "0");
+export const prepByListCM = GetPrepareByCM.slice(0);
 const classes = GetInstructionSource.slice(0);
 const instructors = GetInstructor.slice(0);
 const funding = createSummaryList(GetFundingSource, "FSID", "FundAbbrev");
@@ -38,81 +39,6 @@ export const createOptionList = (dataObj, defaultValue) => {
     }
   });
   return optionList.join("");
-};
-
-const categoryHandler = (titleID, groupID) => {
-  let selectedList = [];
-  //! NOTE: Category IDs are pointing to the newly defined categories.
-  // Categories: 182 = Assessment reports, 57 = Rosters, 46 = Program management
-  //       85 = Exiting and outcome
-  if (["182", "57", "46", "85"].includes(groupID)) {
-    // Category options: Instructional offering, Funding sources
-    if (["47", "86", "46", "85", "207"].includes(titleID)) {
-      $(".category").removeClass("hidden");
-      selectedList = prepByList.filter((record) =>
-        ["1", "2"].includes(record.key)
-      );
-    }
-    // Category options: Instructional offering, Funding sources,
-    // Site, Ref partner, Teacher/Tutor
-    else if (["236", "162", "138", "176", "163", "164"].includes(titleID)) {
-      $(".category").removeClass("hidden");
-      selectedList = prepByList;
-    }
-    // Category options: Instructional offering, Teacher/Tutor
-    else if (["57", "55", "56", "186", "87", "187"].includes(titleID)) {
-      $(".category").removeClass("hidden");
-      selectedList = prepByList.filter((record) =>
-        ["1", "5"].includes(record.key)
-      );
-    }
-    // Report tile with no additional selectors
-    else {
-      return;
-    }
-    $("#report-category").empty().append(createOptionList(selectedList));
-  }
-  // Categories: 242 = Case management reports
-  // Reports ID: 242, 243, 244, 245
-  else if (groupID === "242") {
-    selectedList =
-      titleID === "242"
-        ? // Category options: All Students, Referral partners, Keywords, Date range
-          prepByListCM
-        : // Category options: All Students, Keywords, Date range
-          prepByListCM.filter((record) => record.key !== "2");
-    $(".category").removeClass("hidden");
-    $("#report-category").empty().append(createOptionList(selectedList));
-  }
-  // Category: 133 = Contact hours reports
-  // Reports ID: 133, 185, 140, 169, 170, 236
-  else if (groupID === "133") {
-    // Displays month selector
-    if (["169", "185"].includes(titleID)) {
-      $("#optional-selectors .months").removeClass("hidden");
-      $("#report-months").focus();
-    }
-    // Displays month and week selectors
-    else if (titleID === "170") {
-      $("#optional-selectors .months").removeClass("hidden");
-      $("#optional-selectors .weeks").removeClass("hidden");
-      $("#report-months").focus();
-    }
-    // Contact hours testing summary in group 133
-    // Category options: Instructional offering, Funding sources,
-    // Site, Ref partner, Teacher/Tutor
-    else if (["236", "140"].includes(titleID)) {
-      $(".category").removeClass("hidden");
-      selectedList =
-        titleID === "236"
-          ? prepByList
-          : prepByList.filter((record) => record.key === "1");
-      $(".category").removeClass("hidden");
-      $("#report-category").empty().append(createOptionList(selectedList));
-    } else {
-      return;
-    }
-  }
 };
 
 const criteriaHandler = (selectedCategory, selectedGroup) => {
